review other use of dispatchEvent
